Golang server that responds with Stuff Quiz JSON
Initial Golang server that fetches Stuff Quiz JSON list, retrieves its
riddle Quiz URL, fetches the HTML response of that URL, parses
that HTML for the JavaScript object that contains all the Quiz
information such as questions, answers etc.

This is extremely dirty and just wanted to get it done as fast as
possible to see how plausible it would be. Next steps is to clean this
up.