Created Quiz fetching endpoints
Created 2 endpoints, `/quizzes` and `quizzes/:id`. The former fetches a
list of Quizzes from the stuff JSON API, and the latter fetches the Quiz
JSON from the riddle.com embed for the Quiz information.

The quiz id returned in `/quizzes` is actually the riddle.com quiz id, not
the stuff quiz article id.